foodiebot/requirements.txt
Flask == 2.3.2
numpy == 1.25.0
googlemaps == 4.10.0
retry == 0.9.2
PyJWT == 2.8.0